Moves ledger router into Ledger
